Code clean-up/memory reduction
By moving to a constant dictionary rather than a switch statement, a
small amount of memory is saved, and the code is easier to read. Also
made StatOptions a module, since it only contained static members.

Made a function to obtain the latest sensor history reading rather than
doing it separately for each different sensor.